Moves screenshots to separate JS module
