+++ static/js/publisher/publisher.js
@@ -1,5 +1,5 @@
 import metrics from './metrics/metrics';
 import { selector } from './metrics/filters';
-import market from './market/market';
+import * as market from './market/market';
 
 export { metrics, selector, market };
